added: Component AuthReg & Login
