+++ src/components/AuthReg/Login/index.jsx
@@ -1,0 +1,18 @@
+import AuthReg from "../index";
+
+function Login() {
+  return(
+    <AuthReg title={'Рады видеть!'}>
+      <div className='authreg__input-container'>
+        <span className='authreg__span'>E-mail</span>
+        <input className='authreg__input' type='email' placeholder='pochta@yandex.ru' />
+      </div>
+      <div className='authreg__input-container'>
+        <span className='authreg__span'>Пароль</span>
+        <input className='authreg__input' type='password' />
+      </div>
+    </AuthReg>
+  );
+}
+
+export default Login;
